Move LinkResolver to separate file
